# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "SinkMC Api"
  },
...
  "paths": {
...
    "/v1/resources/{locator}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "sink"
            },
            "required": true,
            "name": "locator",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "A single resource",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resource": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "ownerId": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "category": {
                          "type": "string",
                          "enum": [
                            "Administration",
                            "Tools",
                            "Security",
                            "Social",
                            "Optimization",
                            "Mechanics",
                            "Libraries",
                            "World Generation",
                            "Economy",
                            "Cosmetics",
                            "Minigames"
                          ]
                        },
                        "price": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "createdAt": {
                          "type": "string",
                          "nullable": true,
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "nullable": true,
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "ownerId",
                        "name",
                        "slug",
                        "description",
                        "category",
                        "price",
                        "createdAt",
                        "updatedAt"
                      ],
                      "example": {
                        "id": "_",
                        "ownerId": "_",
                        "name": "My Cool Plugin",
                        "slug": "my-cool-plugin",
                        "description": "A very very neat plugin",
                        "category": "Mechanics",
                        "price": 0,
                        "createdAt": "2026-08-01T16:40:25.000Z",
                        "updatedAt": "2026-08-01T16:40:25.000Z"
                      }
                    }
                  },
                  "required": [
                    "resource"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "NOT_FOUND"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "enum": [
                        "INTERNAL_ERROR"
                      ]
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}